Fix Bug 9958: 'SDP displays double spaces between payload formats as a 0 payload format'
Given an SDP m= line such as this:

    m=audio 29156 RTP/AVP 18  0

SDP will show a media format of G.729 (the 18) and then two G.711 entries: one
for the extra space between the 18 and 0, and one format for the 0. The latter
is correct, but the extra space one isn't.

Technically such an m= line is malformed, since only one space is allowed
between payload formats; but it's definitely not a format of 0.

A similar thing happens in many parts of SDP dissection code. It needs to issue
an expert error and handle it gracefully.

Change-Id: I1f1500489a13a55e03fc8ea14b37d99a019fc449
Reviewed-on: https://code.wireshark.org/review/989
Reviewed-by: Hadriel Kaplan <hadrielk@yahoo.com>
Reviewed-by: Anders Broman <a.broman58@gmail.com>