Add commit-msg hook to tools to spare searching for it
commit-msg script adds change id at end of commit messages.
As this script is needed for Gerrit review system to work 
and every wireshark developer will need to download it from 
somewhere, it will be better if it is downloaded with Wireshark
sources and ready to copy to .git/hooks

Copied from AOSP Gerrit.

Change-Id: Ib3705abfedd2869462eef57690a2f430037f9cc1
Reviewed-on: https://code.wireshark.org/review/1008
Reviewed-by: Michael Mann <mmann78@netscape.net>