ssl: Detect unencrypted heartbeat messages
Records are always unencrypted before the ChangeCipherSpec message. This
patch assumes that conversations without a SSL decoder (i.e. before
calling ssl_change_cipher) are unencrypted. If it turns out that the
contents were encrypted anyway, then there is about 0.8% probability
that the heartbeat message gets recognized wrong
(dissect_ssl3_heartbeat checks if the first byte equals 1 (request) or 2
(response)). Not a big deal, and the advantage that the heartbleed
expert information is triggered overweights that possible mistake.

(Note that ssl_set_master_secret as called by external code will also
invoke ssl_change_cipher, but that should be caught by the if
condition in the line before this hunk.)

An example capture is available at:
https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=9988

Change-Id: I5b14d9e7e8e0c1cd358f3b05c9b84fca1daf1d69
Reviewed-on: https://code.wireshark.org/review/1102
Reviewed-by: Evan Huus <eapache@gmail.com>