Call each item in the Extended IS Reachability CLV an "IS Neighbor".
This makes it a bit more like the (un-extended) Reachability CLVs.

Also, just make the top-level items just a text field, just like the
top-level items for neighbors in the (un-extended) Reachability CLVs.

Rename the system IDs to be "... neighbor ID", not just "neighbor"; the
top-level items hold all the information for the neighbor, and the
system ID for the neighbor is just one of those pieces of information.

Change-Id: I72182dea496f8cfa6048cff3b0d2474df8f31282
Reviewed-on: https://code.wireshark.org/review/1144
Reviewed-by: Guy Harris <guy@alum.mit.edu>