Squelch some warnings:
The dictionary_id member of a spdy_conv_t is only used if we have libz;
put it inside the #ifdef.  That also lets us give it a type of uLong,
from libz, which avoids it being 32 bits when the type returned by
adler32() is 64 bits.  (The *value returned by adler32()* might always
fit in 32 bits, but this is arguably cleaner than throwing a cast at the
problem.)

The third argument to adler32() is a uInt; cast sizeof to uInt to avoid
other 32-bit-vs-64-bit warnings.  (It should have been size_t, but maybe
libz antedated size_t's availability in all the compilers that were used
to compile it.)

The buffer size in spdy_decompress_header_block() is always 16K; just
make it a #define that is *not* size_t, so that we avoid other
32-bit-vs-64-bit warnings.

Use DISSECTOR_ASSERT_NOT_REACHED() for "this can't happen" - it's marked
as "doesn't return", so we don't get "variable is unassigned" warnings
(at least not from the Clang I'm using), and also means we wouldn't get
a crash if it *does* happen (we just get a warning on the console and in
the protocol tree).

Change-Id: I55945b69b7485a02f3f623b21f671ed2915d453d
Reviewed-on: https://code.wireshark.org/review/1162
Reviewed-by: Guy Harris <guy@alum.mit.edu>