Manually apply https://code.wireshark.org/review/1289 to master-1.10:
Fix up get_hfi_and_length():

Treat FT_BYTES and a few others the same as FT_PROTOCOL: allow a zero length
but throw an exception if the offset is already beyond the end of the TVB
(prior to this change it would assert out).  This (when manually applied to
master-1.10) fixes https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=9999 .

Use tvb_ensure_captured_length_remaining() to throw an exception if the offset
is bad.

Change-Id: Ibc61cc9ac331b669c688611ee9f23d36479378ef
Reviewed-on: https://code.wireshark.org/review/1290
Reviewed-by: Alexis La Goutte <alexis.lagoutte@gmail.com>
Reviewed-by: Evan Huus <eapache@gmail.com>