Fix procedure of finding packages
Before this change the code using make package was just showing
a string ${PACKAGE_VAR}_FOUND instead of showing if the
${PACKAGE_VAR}_FOUND variable was set true by find_package function.
Now the user will be informed about which packages were not found
and an error will be raised, as packages sought are the ones that
are selected for the build.

This change also ads fixes to some improper PACKAGE_VAR
variables that prevented find_package from finding
proper packages.

Change-Id: Ic20bc50ace65bab385059a7430909f95367520a3
Reviewed-on: https://code.wireshark.org/review/1410
Reviewed-by: Evan Huus <eapache@gmail.com>
Reviewed-by: Alexis La Goutte <alexis.lagoutte@gmail.com>