Initial Ceph Dissector.
This is the first version of a Ceph dissector.  It is not complete but
is far enough along to be helpful to many people working with Ceph.

Currently the dissector can fully dissect the Ceph protocol and has
support for full dissection of most common messages.  For the other
messages for which full dissection is not available their metadata is
parsed and shown along with the raw data of the different message
sections.

Change-Id: Ic7917a3d01148c6fe2f9ea2c13ecd09ecc06c2d7
Reviewed-on: https://code.wireshark.org/review/1889
Reviewed-by: Bill Meier <wmeier@newsguy.com>