Be less aggressive about rebuilding sminmpec.c.
Instead of forcing developers to generate sminmpec.c (which will have
different results depending on the presence or absence of a working
Internet connection) add sminmpec.c back to the repository. I'll add
it to the weekly update-numbers script so that it will be updated at
the same time as manuf, services, enterprise-numbers, and usb.c.

Change the Autotools, CMake, and Nmake sminmpec.c target name to
"update-sminmpec".

Remove the mtime check from make-sminmpec.pl. Update enterprise-numbers
and sminmpec.c while we're here.

Tested with an in-tree Autotools build and an out-of-tree CMake build.

Change-Id: Iecc332ce2731e3e98ab0205a56c78807e599a026
Reviewed-on: https://code.wireshark.org/review/1516
Reviewed-by: Evan Huus <eapache@gmail.com>
Reviewed-by: Alexis La Goutte <alexis.lagoutte@gmail.com>
Tested-by: Alexis La Goutte <alexis.lagoutte@gmail.com>
(cherry picked from commit 3f247a5cee61d8f2139ade94cd9840b100337b05 with manual fixups)
Reviewed-on: https://code.wireshark.org/review/1577
Reviewed-by: Gerald Combs <gerald@wireshark.org>