This fix now allows SASL/GSS-API with integrity only to be properly dissected in that we now properly indicate the checksum and provide the wrapped token to the LDAP dissector to dissect.
There is still a problem in that the wrong parts of the SASL packet can be
indicated when users click on some parts.

I am committing it for review and because it is much better than the current
code.

Bug:9398
Change-Id: I552fc249db26835b79abfc8793438a95f221663a
Signed-off-by: Richard Sharpe <realrichardsharpe@gmail.com>
Reviewed-on: https://code.wireshark.org/review/1905
Reviewed-by: Evan Huus <eapache@gmail.com>
Reviewed-by: Alexis La Goutte <alexis.lagoutte@gmail.com>
(cherry picked from commit 09b5361302725cc83b7c8d76282441d031050c38)
Reviewed-on: https://code.wireshark.org/review/2076
Reviewed-by: Michael Mann <mmann78@netscape.net>