EIGRP Sequence TLV Dissector Patch
In the EIGRP packet dissector, the dissector routine for the Sequence TLV
dissected the TLV only up to the first address in the list. However, the
Sequence TLV contains a variably sized list of addresses. This patch
modifies the routine so that it processes the entire TLV, not just the
first address in the contained list.

Also, in the dissect_eigrp(), replaced calls to tvb_new_subset() with
the reported length set to -1 with the call to tvb_new_subset_length().
TLVs always carry information about their length. And this time,
correct truly ALL calls in the switch{} section.

Sample packet capture is available in BugZilla.

Bug: 10156
Change-Id: Idaaf182c05bcf799f770f23a2ce2b1e05a3d569a
Reviewed-on: https://code.wireshark.org/review/1911
Reviewed-by: Pascal Quantin <pascal.quantin@gmail.com>

(cherry picked from commit 75e53a165c1167607f49962919df2e946e6476cf)

Change-Id: Ia8e2c53bf227b0028691f1a95f6947921133a19e
Reviewed-on: https://code.wireshark.org/review/1923
Reviewed-by: Pascal Quantin <pascal.quantin@gmail.com>