Sync .deb packaging from Debian
The sync makes CMake the build system for the .deb package
and starts providing wireshark-qt in the wireshark-qt package.
The package structure, i.e. the libraries are shipped in separate
packages is also sync-ed.

Wireshark-qt uses the Qt 4 libraries, but it is easy to switch
it to Qt 5.

Change-Id: I849d18bdb8ca6ebf4072cf1d73d749080ac5dac2
Reviewed-on: https://code.wireshark.org/review/1986
Reviewed-by: Gerald Combs <gerald@wireshark.org>
Reviewed-by: Balint Reczey <balint@balintreczey.hu>
Tested-by: Balint Reczey <balint@balintreczey.hu>
(cherry-picked from commit 97e1491aa8f418a7f5bd7fdd92561d2403c187f6)
Reviewed-on: https://code.wireshark.org/review/2058
Reviewed-by: Alexis La Goutte <alexis.lagoutte@gmail.com>