Revert "Fix the fuzz failure reported in https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=8940"
While the bugfix makes sense, it requires a huge amount of work to get the messages decoded again (backport of gae61fe0 and ge8c305f; see http://ask.wireshark.org/questions/33502/why-cannot-fully-decode-nas-field-on-s1ap-protocol).
As I do not feel courageous enough to backport everything, let's go back to the old code for this soon to be dead branch.

This reverts commit 9f46f19b276ed3b166ef7931b971caa7c4eaa082.

Change-Id: I3251044d0ed7607badce3679a1af89d15e00e458
Reviewed-on: https://code.wireshark.org/review/2020
Reviewed-by: Michael Mann <mmann78@netscape.net>